requirements.txt
Flask
tensorflow
numpy
Pillow
requests
python-dotenv

